requirements.txt
bs4==0.0.1
lxml==4.8.0
python-docx==0.8.11
requests==2.27.1